Quick refactor
