+++ nats-test/src/events/ticket-created-listener.ts
@@ -1,0 +1,15 @@
+
+import { Listener } from './base-listeners';
+import { Message } from 'node-nats-streaming';
+
+export class TicketCreatedListener extends Listener {
+    subject = 'ticket:created';
+    queueGroupName = 'payment-service';
+
+    onMessage(data: any, msg: Message): void {
+        console.log('Event data!', data);
+
+        msg.ack();
+    }
+
+}
